Revert "CCIP-2791 Fixes to the snoozed roots - more generic way of passing fi…"
This reverts commit ac3d4790d716020b40ac93f3faaab0d2b2ef979a.
