+++ pkg/types/ccip/offramp.go
@@ -50,6 +50,7 @@ type ExecutionStateChangedWithTxMeta struct {
 
 type ExecutionStateChanged struct {
 	SequenceNumber uint64
+	Finalized      bool
 }
 
 type TokenBucketRateLimit struct {
